Fixed the type support issue

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdint.h>
 
 /**
  * main - prints the even fibonacci numbers that are less than 4000000
@@ -7,18 +8,20 @@
  */
 int main(void)
 {
-	long int maxValue = 4000000;
-	long int num1 = 0;
-	long int num2 = 1;
-	long sum;
+	int64_t maxValue = 4000000;
+	int64_t num1 = 0;
+	int64_t num2 = 1;
+	int64_t sum;
 
 	while (num2 < maxValue)
 	{
-		int temp = num1;
+		int64_t temp = num1;
+
 		num1 = num2;
-		num2+= temp;
+		num2 += temp;
+
 		if (num2 % 2 == 0)
-			sum+= num2;
+			sum += num2;
 	}
 	printf("%ld\n", sum);
